Refactor Device into its own file
